final fixes

+++ src/main/java/com/example/topoftops/controller/command/impl/DeleteItemCommand.java
@@ -3,7 +3,9 @@ package com.example.topoftops.controller.command.impl;
 import com.example.topoftops.controller.command.Command;
 import com.example.topoftops.controller.command.PagePath;
 import com.example.topoftops.controller.command.Router;
+import com.example.topoftops.entity.User;
 import com.example.topoftops.entity.Item;
+import com.example.topoftops.entity.Role;
 import com.example.topoftops.entity.Top;
 import com.example.topoftops.exception.ServiceException;
 import com.example.topoftops.model.service.ItemService;
@@ -18,6 +20,12 @@ import java.util.ArrayList;
 
 import static com.example.topoftops.controller.command.RequestParam.*;
 
+/**
+ * The command is responsible for delete item
+ *
+ * @author Ilya Tsvetkov
+ * @see Command
+ */
 public class DeleteItemCommand implements Command {
     private static final Logger logger = LogManager.getLogger();
     private static final String ATTRIBUTE_TOP = "top";
@@ -35,16 +43,23 @@ public class DeleteItemCommand implements Command {
         String page = ConfigurationManager.getProperty(PagePath.TOP);
         long topId = Long.parseLong(request.getParameter(PARAM_NAME_TOP));
         long itemId = Long.parseLong(request.getParameter(PARAM_NAME_ITEM));
+        User user = (User) request.getSession().getAttribute("user");
         Top top = null;
         try {
             top = (Top) topService.findById(topId).get();
-            itemService.delete(itemId);
-            ArrayList<Item> items = itemService.find(topId);//fixme idk how to do it right
-            top.setItems(items);
+            if (user.getId() == top.getUser() || user.getRole() == Role.ADMIN.ordinal()) {
+                itemService.delete(itemId);
+                ArrayList<Item> items = itemService.findItems(topId);
+                top.setItems(items);
+            } else {
+                logger.log(Level.ERROR, "user tried delete item to stranger top");
+                page = ConfigurationManager.getProperty(PagePath.INDEX);
+            }
         } catch (ServiceException e) {
-            logger.log(Level.ERROR,e);
+            logger.log(Level.ERROR, e);
+            page = ConfigurationManager.getProperty(PagePath.INDEX);
         }
-        request.setAttribute(ATTRIBUTE_TOP,top);
+        request.setAttribute(ATTRIBUTE_TOP, top);
         router = new Router(page, Router.RouteType.FORWARD);
         return router;
     }
